Add window with backend is OpenGL framework GLFW and add event callback

+++ GGEngine/GGEngine_Core/Event/applicationevent.h
@@ -8,11 +8,11 @@ namespace GGEngine {
 
 class GGE_API WindowResizeEvent : public Event {
 public:
-    WindowResizeEvent(float width, float height)
+    WindowResizeEvent(unsigned int width, unsigned int height)
         :m_Width(width), m_Height(height) {}
 
-    inline float GetWidth() const { return m_Width; }
-    inline float GetHeight() const { return m_Height; }
+    inline unsigned int GetWidth() const { return m_Width; }
+    inline unsigned int GetHeight() const { return m_Height; }
 
     std::string ToString() const override
     {
@@ -25,7 +25,7 @@ public:
     EVENT_CATEGORY(EventCategoryApplication)
 
 private:
-    float m_Width, m_Height;
+    unsigned int m_Width, m_Height;
 };
 
 class GGE_API WindowCloseEvent : public Event {
